# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Stealth Console API
  version: 0.1.0
  description: Control-plane API for Stealth Console. Browser sessions are opaque HttpOnly cookies and are never returned in JSON.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/projects/{projectID}/usage:
    get:
      operationId: getProjectUsage
      description: Return a live aggregate snapshot of tenant-owned project resources plus rolling 30-day API and Functions metering counters.
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }]
      responses:
        '200': { description: Project usage snapshot, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectUsageResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/projects/{projectID}/usage/metering:
    get:
      operationId: getProjectUsageMetering
      description: Return durable daily API request, egress, Function invocation, failure, and compute-time buckets for a bounded inclusive calendar window. Empty days are omitted; totals cover the returned buckets.
      security: [{ sessionCookie: [] }]
      parameters:
        - { $ref: '#/components/parameters/ProjectID' }
        - { name: from, in: query, required: false, description: Inclusive UTC start date. Defaults to 29 days before today., schema: { type: string, format: date } }
        - { name: to, in: query, required: false, description: Inclusive UTC end date. Defaults to today., schema: { type: string, format: date } }
        - { name: format, in: query, required: false, description: Response format. CSV exports include the exact daily buckets followed by a TOTAL row., schema: { type: string, enum: [json, csv], default: json } }
      responses:
        '200': { description: Daily project usage metering, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectUsageMeteringResponse' } }, text/csv: { schema: { type: string, format: binary } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: stealth_session }
    projectAppCookie: { type: apiKey, in: cookie, name: stealth_app_<project-id-without-hyphens>, description: Project-specific opaque HttpOnly cookie. The actual name is derived from the path projectID; it is never the Console stealth_session cookie. }
    projectApiKey: { type: apiKey, in: header, name: X-Stealth-Key, description: Project-bound server-to-server secret. Supported scopes include users, databases, storage, functions, sites, webhooks, realtime.read, and messaging read/write pairs; never send this header through the Console browser bridge. }
  parameters:
...
    ProjectID: { name: projectID, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
...
    Unauthorized: { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    Forbidden: { description: Insufficient organization role, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    InternalError: { description: Internal server error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    ValidationError: { description: Validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
  schemas:
...
    ErrorEnvelope: { type: object, additionalProperties: false, required: [error], properties: { error: { type: object, additionalProperties: false, required: [code, message], properties: { code: { type: string }, message: { type: string } } } } }
...
    ProjectUsage: { type: object, additionalProperties: false, required: [project_id, captured_at, application_users, database_count, database_table_count, database_row_count, storage_file_count, storage_bytes, storage_quota_bytes, function_count, function_artifact_bytes, function_quota_bytes, site_count, site_artifact_bytes, site_reserved_bytes, site_quota_bytes, realtime_event_count, webhook_delivery_count_7d, api_request_count_30d, api_egress_bytes_30d, function_invocation_count_30d, function_failure_count_30d, function_compute_ms_30d], properties: { project_id: { type: string, format: uuid }, captured_at: { type: string, format: date-time }, application_users: { type: integer, format: int64, minimum: 0 }, database_count: { type: integer, format: int64, minimum: 0 }, database_table_count: { type: integer, format: int64, minimum: 0 }, database_row_count: { type: integer, format: int64, minimum: 0 }, storage_file_count: { type: integer, format: int64, minimum: 0 }, storage_bytes: { type: integer, format: int64, minimum: 0 }, storage_quota_bytes: { type: integer, format: int64, minimum: 0 }, function_count: { type: integer, format: int64, minimum: 0 }, function_artifact_bytes: { type: integer, format: int64, minimum: 0 }, function_quota_bytes: { type: integer, format: int64, minimum: 0 }, site_count: { type: integer, format: int64, minimum: 0 }, site_artifact_bytes: { type: integer, format: int64, minimum: 0 }, site_reserved_bytes: { type: integer, format: int64, minimum: 0 }, site_quota_bytes: { type: integer, format: int64, minimum: 0 }, realtime_event_count: { type: integer, format: int64, minimum: 0 }, webhook_delivery_count_7d: { type: integer, format: int64, minimum: 0 }, api_request_count_30d: { type: integer, format: int64, minimum: 0 }, api_egress_bytes_30d: { type: integer, format: int64, minimum: 0 }, function_invocation_count_30d: { type: integer, format: int64, minimum: 0 }, function_failure_count_30d: { type: integer, format: int64, minimum: 0 }, function_compute_ms_30d: { type: integer, format: int64, minimum: 0 } } }
    ProjectUsageDay: { type: object, additionalProperties: false, required: [date, api_request_count, api_egress_bytes, function_invocation_count, function_failure_count, function_compute_ms], properties: { date: { type: string, format: date }, api_request_count: { type: integer, format: int64, minimum: 0 }, api_egress_bytes: { type: integer, format: int64, minimum: 0 }, function_invocation_count: { type: integer, format: int64, minimum: 0 }, function_failure_count: { type: integer, format: int64, minimum: 0 }, function_compute_ms: { type: integer, format: int64, minimum: 0 } } }
    ProjectUsageMetering: { type: object, additionalProperties: false, required: [project_id, from, to, days, totals], properties: { project_id: { type: string, format: uuid }, from: { type: string, format: date }, to: { type: string, format: date }, days: { type: array, maxItems: 367, items: { $ref: '#/components/schemas/ProjectUsageDay' } }, totals: { $ref: '#/components/schemas/ProjectUsageDay' } } }
    ProjectUsageMeteringResponse: { type: object, additionalProperties: false, required: [metering], properties: { metering: { $ref: '#/components/schemas/ProjectUsageMetering' } } }
    ProjectUsageResponse: { type: object, additionalProperties: false, required: [usage], properties: { usage: { $ref: '#/components/schemas/ProjectUsage' } } }
...